requirements.txt
python-telegram-bot==20.6
python-dotenv==1.0.0
web3==6.10.0
aiohttp==3.8.5
motor==3.3.1
pymongo==4.5.0
pandas==2.1.1
matplotlib==3.8.0
